# reuse definition and repetition buffers