--- a/page.go
+++ b/page.go
@@ -376,10 +376,10 @@ func errPageBoundsOutOfRange(i, j, n int64) error {
 type optionalPage struct {
 	base               Page
 	maxDefinitionLevel byte
-	definitionLevels   bufferRef
+	definitionLevels   []byte
 }
 
-func newOptionalPage(base Page, maxDefinitionLevel byte, definitionLevels bufferRef) *optionalPage {
+func newOptionalPage(base Page, maxDefinitionLevel byte, definitionLevels []byte) *optionalPage {
 	return &optionalPage{
 		base:               base,
 		maxDefinitionLevel: maxDefinitionLevel,
@@ -393,21 +393,21 @@ func (page *optionalPage) Column() int { return page.base.Column() }
 
 func (page *optionalPage) Dictionary() Dictionary { return page.base.Dictionary() }
 
-func (page *optionalPage) NumRows() int64 { return int64(page.definitionLevels.len) }
+func (page *optionalPage) NumRows() int64 { return int64(len(page.definitionLevels)) }
 
-func (page *optionalPage) NumValues() int64 { return int64(page.definitionLevels.len) }
+func (page *optionalPage) NumValues() int64 { return int64(len(page.definitionLevels)) }
 
 func (page *optionalPage) NumNulls() int64 {
-	return int64(countLevelsNotEqual(page.definitionLevels.data(), page.maxDefinitionLevel))
+	return int64(countLevelsNotEqual(page.definitionLevels, page.maxDefinitionLevel))
 }
 
 func (page *optionalPage) Bounds() (min, max Value, ok bool) { return page.base.Bounds() }
 
-func (page *optionalPage) Size() int64 { return page.base.Size() + int64(page.definitionLevels.len) }
+func (page *optionalPage) Size() int64 { return page.base.Size() + int64(len(page.definitionLevels)) }
 
 func (page *optionalPage) RepetitionLevels() []byte { return nil }
 
-func (page *optionalPage) DefinitionLevels() []byte { return page.definitionLevels.data() }
+func (page *optionalPage) DefinitionLevels() []byte { return page.definitionLevels }
 
 func (page *optionalPage) Data() encoding.Values { return page.base.Data() }
 
@@ -422,18 +422,19 @@ func (page *optionalPage) Clone() Page {
 	return newOptionalPage(
 		page.base.Clone(),
 		page.maxDefinitionLevel,
-		page.definitionLevels.clone(),
+		copyBytes(page.definitionLevels),
 	)
 }
 
 func (page *optionalPage) Slice(i, j int64) Page {
-	definitionLevels := page.definitionLevels.data()
-	numNulls1 := int64(countLevelsNotEqual(definitionLevels[:i], page.maxDefinitionLevel))
-	numNulls2 := int64(countLevelsNotEqual(definitionLevels[i:j], page.maxDefinitionLevel))
+	maxDefinitionLevel := page.maxDefinitionLevel
+	definitionLevels := page.definitionLevels
+	numNulls1 := int64(countLevelsNotEqual(definitionLevels[:i], maxDefinitionLevel))
+	numNulls2 := int64(countLevelsNotEqual(definitionLevels[i:j], maxDefinitionLevel))
 	return newOptionalPage(
 		page.base.Slice(i-numNulls1, j-(numNulls1+numNulls2)),
-		page.maxDefinitionLevel,
-		page.definitionLevels.slice(int(i), int(j)),
+		maxDefinitionLevel,
+		definitionLevels[i:j:j],
 	)
 }
 
@@ -441,11 +442,11 @@ type repeatedPage struct {
 	base               Page
 	maxRepetitionLevel byte
 	maxDefinitionLevel byte
-	definitionLevels   bufferRef
-	repetitionLevels   bufferRef
+	definitionLevels   []byte
+	repetitionLevels   []byte
 }
 
-func newRepeatedPage(base Page, maxRepetitionLevel, maxDefinitionLevel byte, repetitionLevels, definitionLevels bufferRef) *repeatedPage {
+func newRepeatedPage(base Page, maxRepetitionLevel, maxDefinitionLevel byte, repetitionLevels, definitionLevels []byte) *repeatedPage {
 	return &repeatedPage{
 		base:               base,
 		maxRepetitionLevel: maxRepetitionLevel,
@@ -461,27 +462,23 @@ func (page *repeatedPage) Column() int { return page.base.Column() }
 
 func (page *repeatedPage) Dictionary() Dictionary { return page.base.Dictionary() }
 
-func (page *repeatedPage) NumRows() int64 {
-	return int64(countLevelsEqual(page.repetitionLevels.data(), 0))
-}
+func (page *repeatedPage) NumRows() int64 { return int64(countLevelsEqual(page.repetitionLevels, 0)) }
 
-func (page *repeatedPage) NumValues() int64 {
-	return int64(page.definitionLevels.len)
-}
+func (page *repeatedPage) NumValues() int64 { return int64(len(page.definitionLevels)) }
 
 func (page *repeatedPage) NumNulls() int64 {
-	return int64(countLevelsNotEqual(page.definitionLevels.data(), page.maxDefinitionLevel))
+	return int64(countLevelsNotEqual(page.definitionLevels, page.maxDefinitionLevel))
 }
 
 func (page *repeatedPage) Bounds() (min, max Value, ok bool) { return page.base.Bounds() }
 
 func (page *repeatedPage) Size() int64 {
-	return int64(page.repetitionLevels.len) + int64(page.definitionLevels.len) + page.base.Size()
+	return int64(len(page.repetitionLevels)) + int64(len(page.definitionLevels)) + page.base.Size()
 }
 
-func (page *repeatedPage) RepetitionLevels() []byte { return page.repetitionLevels.data() }
+func (page *repeatedPage) RepetitionLevels() []byte { return page.repetitionLevels }
 
-func (page *repeatedPage) DefinitionLevels() []byte { return page.definitionLevels.data() }
+func (page *repeatedPage) DefinitionLevels() []byte { return page.definitionLevels }
 
 func (page *repeatedPage) Data() encoding.Values { return page.base.Data() }
 
@@ -497,8 +494,8 @@ func (page *repeatedPage) Clone() Page {
 		page.base.Clone(),
 		page.maxRepetitionLevel,
 		page.maxDefinitionLevel,
-		page.repetitionLevels.clone(),
-		page.definitionLevels.clone(),
+		copyBytes(page.repetitionLevels),
+		copyBytes(page.definitionLevels),
 	)
 }
 
@@ -514,7 +511,11 @@ func (page *repeatedPage) Slice(i, j int64) Page {
 		panic(errPageBoundsOutOfRange(i, j, numRows))
 	}
 
-	repetitionLevels := page.repetitionLevels.data()
+	maxRepetitionLevel := page.maxRepetitionLevel
+	maxDefinitionLevel := page.maxDefinitionLevel
+	repetitionLevels := page.repetitionLevels
+	definitionLevels := page.definitionLevels
+
 	rowIndex0 := 0
 	rowIndex1 := len(repetitionLevels)
 	rowIndex2 := len(repetitionLevels)
@@ -539,19 +540,18 @@ func (page *repeatedPage) Slice(i, j int64) Page {
 		}
 	}
 
-	definitionLevels := page.definitionLevels.data()
-	numNulls1 := countLevelsNotEqual(definitionLevels[:rowIndex1], page.maxDefinitionLevel)
-	numNulls2 := countLevelsNotEqual(definitionLevels[rowIndex1:rowIndex2], page.maxDefinitionLevel)
+	numNulls1 := countLevelsNotEqual(definitionLevels[:rowIndex1], maxDefinitionLevel)
+	numNulls2 := countLevelsNotEqual(definitionLevels[rowIndex1:rowIndex2], maxDefinitionLevel)
 
 	i = int64(rowIndex1 - numNulls1)
 	j = int64(rowIndex2 - (numNulls1 + numNulls2))
 
 	return newRepeatedPage(
 		page.base.Slice(i, j),
-		page.maxRepetitionLevel,
-		page.maxDefinitionLevel,
-		page.repetitionLevels.slice(int(rowIndex1), int(rowIndex2)),
-		page.definitionLevels.slice(int(rowIndex1), int(rowIndex2)),
+		maxRepetitionLevel,
+		maxDefinitionLevel,
+		repetitionLevels[rowIndex1:rowIndex2:rowIndex2],
+		definitionLevels[rowIndex1:rowIndex2:rowIndex2],
 	)
 }